requirements.txt
streamlit>=1.33.0
fpdf>=1.7.2
pandas>=2.2.2
python-dotenv>=1.0.1
langchain>=0.1.14
langchain-groq>=0.0.6
